# doc/api.yaml
openapi: 3.0.0  
info:
  title: Simple API overview
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Test server
  - url: http://waterfall.wwyf.top
    description: Production server
paths:
  # /apis:
  # /apis/user:
  # /apis/user/login:
  # /apis/user/register:
  # /apis/user/logout:
  # /apis/user/{userId}:
  /apis/order/mainOrder:
    get:
      tags: 
        - 母订单相关
      summary: 获取所有主订单
      description: 获得所有/前limit个订单，并且获得每个订单的内部详细字段。
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/skipParam'
      responses:
        200:
          description: 成功获取并返回订单信息
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/mainOrder_view'
                          
    post:
      tags: 
        - 母订单相关
      summary: 提交新母订单
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name: 
                  description: 订单名称
                  type: string
                  example: 鲜天下实例订单名称
                summary :
                  description: 订单概述
                  type: string
                  example: 鲜天下示例订单概述，我司需要大批量的i7-9900K。
                deadline:
                  description : 订单截止日期
                  type: string
                  format: datatime
                  example: "2019-02-05"
                address:
                  description: 订单收货地址
                  type: string
                  example: "广州市广州大学城外环东路132号"
                quantity :
                  description: 货物数量
                  type: integer
                  example: 50
                price :
                  description: 订单单价
                  type: integer
                  example: 100
                createuser:
                  description: 当前用户的id
                  type: integer
                  example: 1
                comments :
                  description: 订单备注
                  type: string
                  example: 空备注
                phone :
                  description: 联系方式
                  type: string
                  example: 13538383838
      responses:
        200:
          $ref: '#/components/responses/GeneralResponseWithId'
          

  /apis/order/mainOrder/{mainOrderId}:
    parameters:
      - $ref: '#/components/parameters/mainOrderIdParam'
    get:
      tags: 
        - 母订单相关
      summary: 获取指定母订单信息
      description: 获取指定母订单信息
      responses:
        200:
          description: 成功获取并返回指定订单信息
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/mainOrder_view'
                      msg:
                        type: string
                        example: "successful"
    post:
      tags: 
        - 母订单相关
      description: 编辑指定母订单信息，只修改已填写字段
      summary: 编辑指定母订单
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name: 
                  description: 订单名称
                  type: string
                  example: 鲜天下实例订单名称
                summary :
                  description: 订单概述
                  type: string
                  example: 鲜天下示例订单概述，我司需要大批量的i7-9900K。
                deadline:
                  description : 订单截止日期
                  type: string
                  format: datatime
                  example: "2019-02-05"
                address:
                  description: 订单收货地址
                  type: string
                  example: "广州市广州大学城外环东路132号"
                quantity :
                  description: 货物数量
                  type: integer
                  example: 50
                price :
                  description: 订单单价
                  type: integer
                  example: 100
                comments :
                  description: 订单备注
                  type: string
                  example: 空备注
                phone :
                  description: 联系方式
                  type: string
                  example: 13538383838
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'
  /apis/order/mainOrder/{mainOrderId}/finish:
    parameters:
      - $ref: '#/components/parameters/mainOrderIdParam'
    post:
      tags: 
        - 母订单相关
      summary: 完成指定订单
      description: 讲该订单的状态修改为完成状态
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'

  /apis/order/mainOrder/{mainOrderId}/cancel:
    parameters:
      - $ref: '#/components/parameters/mainOrderIdParam'
    post:
      tags: 
        - 母订单相关
      summary: 取消指定订单
      description: 讲该订单的状态修改为取消状态
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'
  /apis/order/subOrder:
    get:
      tags: 
        - 子订单相关
      summary: 获取所有子订单
      description: 获得所有/前limit个订单，并且获得每个订单的内部详细字段。
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/skipParam'
      responses:
        200:
          description: 成功获取并返回子订单信息
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/subOrder_view'
    post:
      tags: 
        - 子订单相关
      summary: 提交新子订单
      requestBody:
        content:
          application/json:
            schema:
              properties:
                mainOrderId:
                  type: integer
                  description: 与该子订单对应的母订单id
                  example: 1
                quantity:
                  title: 子订单所能提供货物数量
                  type: integer
                  example : 3
                createuser:
                  title: 提供该子订单的用户id
                  type: integer
                  format: int64
                  example: 8
                comments :
                  description: 订单备注
                  type: string
                  example: 空备注
                phone :
                  description: 联系方式
                  type: string
                  example: 13538383838
      responses:
        200:
          $ref: '#/components/responses/GeneralResponseWithId'
          
  /apis/order/subOrder/{subOrderId}:
    parameters:
      - $ref: '#/components/parameters/subOrderIdParam'
    get:
      tags: 
        - 子订单相关
      summary: 获取指定子订单信息
      description: 获取指定子订单信息
      responses:
        200:
          description: 成功获取并返回指定子订单信息
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/subOrder_view'
                      msg:
                        type: string
                        example: "successful"
    post:
      tags: 
        - 子订单相关
      description: 编辑指定子订单信息，只修改已填写字段
      summary: 编辑指定子订单
      requestBody:
        content:
          application/json:
            schema:
              properties:
                quantity:
                  title: 子订单所能提供货物数量
                  type: integer
                  example : 3
                comments :
                  description: 订单备注
                  type: string
                  example: 空备注
                phone :
                  description: 联系方式
                  type: string
                  example: 13538383838
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'
  /apis/order/subOrder/{subOrderId}/finish:
    parameters:
      - $ref: '#/components/parameters/subOrderIdParam'
    post:
      tags: 
        - 子订单相关
      summary: 完成指定子订单
      description: 讲该子订单的状态修改为完成状态
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'

  /apis/order/subOrder/{subOrderId}/cancel:
    parameters:
      - $ref: '#/components/parameters/subOrderIdParam'
    post:
      tags: 
        - 子订单相关
      summary: 取消指定子订单
      description: 讲该子订单的状态修改为取消状态
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'

  /wallet/{userId}:
    parameters:
      - $ref: '#/components/parameters/userIdParam'
    get:
      tags:
        - 个人钱包
      description: 获取指定用户钱包余额
      summary: 获取指定用户钱包余额
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: integer
                    description: 错误码，0为正常
                    example: 0
                  data:
                    type: object
                    properties:
                      balance:
                        title: 指定用户的余额
                        type: integer
                        example: 89238
                      msg:
                        type: string
                        example: "已提交成功"
                        
  /wallet/{userId}/deposit:
    parameters:
      - $ref: '#/components/parameters/userIdParam'
    post:
      tags:
        - 个人钱包
      summary : 充值接口
      description: 充值接口
      requestBody: 
        content:
          application/json:
            schema:
              properties:
                amount:
                  description: 充值金额
                  type: number
                  example: 34234
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'
  /wallet/{userId}/withdraw:
    parameters:
      - $ref: '#/components/parameters/userIdParam'
    post:
      tags:
        - 个人钱包
      summary : 提现接口
      description: 提现接口
      requestBody: 
        content:
          application/json:
            schema:
              properties:
                amount:
                  description: 提现金额
                  type: number
                  example: 34234
      responses:
        200:
          $ref: '#/components/responses/GeneralResponse'

components:
  schemas:
    Users:
      type: object
      properties:
        ID:
          title : 在数据库中用户的唯一ID
          type: integer
          format : int64
        username:
          title : 用户名
          type: string
        password:
          title : 哈希后的用户密码
          type: string
        email:
          title: 用户电子邮箱
          type: string
          default: null
        phone:
          title: 用户手机号码
          type: string
        balance:
          title: 用户余额
          type: integer
          format: int64
          default: 0
        usertype:
          title: 用户类型（0为管理员，1为订购商，2为微供应商）
          type: integer
          enum:
            - 0
            - 1
            - 2
    Orders:
      type: object
      properties:
        ID:
          title : 在数据库中订单的唯一ID
          type: integer
          format : int64
        name:
          title: 订单名称
          type: string
        summary:
          title: 订单概要/描述
          type: string    
        createdate:
          title: 订单创建日期时间
          type: string
          format: datetime
        deadline:
          title: 订单截止时间
          type: string
          format: datetime
        address:
          title: 订单收货地址
          type: string
        quantity:
          title: 订单货物数量
          type: integer
        price:
          title: 单件货物数量
          type: integer
        totalprice:
          title: 订单总价格
          type: integer
        createuser:
          title: 下达该订单的用户id
          type: integer
          format: int64
        comments:
          title: 订单备注
          type: string
        phone:
          title: 联系方式（即手机号码）
          type: string
        status:
          title: 订单状态(TODO)
          type: integer
          enum:
            - 0
            - 1
            - 2
        progress:
          title: 订单进度
          type: integer
          maximum: 100
          minimum   : 0
    subOrders:
      properties:
        ID:
          description: 在数据库中子订单的唯一ID
          type: integer
          format : int64
        mainOrderId:
          type: integer
          description: 与该子订单对应的母订单id
        createdate:
          description: 订单创建日期时间
          type: string
          format: datetime
        supplier:
          description: 该子订单供应商的id
          type: string
        quantity:
          description: 订单货物数量
          type: integer
        comments:
          description: 订单备注
          type: string
        phone:
          description: 联系方式（即手机号码）
          type: string
        status:
          description: 订单状态(TODO)
          type: integer
          enum:
            - 0
            - 1
            - 2
    mainOrder_view:
      type: object
      properties:
        address:
          title: 订单收货地址
          type: string
        comments:
          description: 订单备注
          type: string
        createdate:
          description: 订单创建日期时间
          type: string
          format: datetime
        createuser:
          description: 下达该订单的用户id
          type: integer
          format: int64
        deadline:
          type: string
          format: datetime
          description: 母订单截止日期
          example: "2018-01-09 11:11:11"
        id:
          type: string
          description: 母订单id
        name:
          type: string
          description: 母订单标题
        phone:
          description: 联系方式（即手机号码）
          type: string
        price:
          description: 单件货物价格
          type: integer
          example : 534
        progress:
          title: 订单进度
          type: integer
          maximum: 100
          minimum   : 0
        quantity:
          description: 母订单所需货物数量
          type: integer
          example : 3
        status:
          description: 订单状态(TODO)
          type: integer
          enum:
            - 0
            - 1
            - 2
        summary:
          type: string
          description: 母订单概要
        totalprice:
          description: 订单总价格
          type: integer
          example: 243423


    subOrder_view:
      type: object
      properties:
        comments:
          description: 订单备注
          type: string
        createdate:
          type: string
          format: datetime
          description: 子订单创建日期时间
          example: "2018-01-02 11:11:11"
        createuser:
          description: 下达该子订单的用户id，也就是该子订单对应微供应商
          type: integer
          format: int64
        id:
          type: string
          description: 子订单id
        mainOrderId:
          type: string
          description: 与该子订单对应的母订单id
        phone:
          description: 联系方式（即手机号码）
          type: string
        quantity:
          description: 子订单所能提供货物数量
          type: integer
          example : 3

        status:
          description: 订单状态(1：正在进行中 3：已完成 4：已取消)
          type: integer
          enum:
            - 0
            - 1
            - 2
            - 3
            - 4

        
  parameters:
    skipParam:
      name: skip
      in: query
      description: number of items to skip
      # required: true
      schema:
        type: integer
        format: int32
    limitParam:
      name: limit
      in: query
      description: max records to return
      # required: true
      schema:
        type: integer
        format: int32
    mainOrderIdParam:
      name : mainOrderId
      description: 母订单在数据库中的id
      in : path
      required : true
      schema:
        title: 订单id
        type: integer
    subOrderIdParam:
      name : subOrderId
      description: 子订单在数据库中的id
      in : path
      required : true
      schema:
        title: 订单id
        type: integer
    userIdParam:
      name : userId
      description: 用户在数据库中的id
      in : path
      required : true
      schema:
        title: 用户id
        type: integer
  responses:
    GeneralResponseWithId:
      description: General response
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                description: 错误码，0为正常
                example: 0
              data:
                type: object
                properties:
                  msg:
                    description: 提交相关信息
                    type: string
                    example: "已提交成功"
                  id:
                    description: 所提交订单在数据库中的id
                    type: integer
                    example: 89238
    GeneralResponse:
      description: General response
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                description: 错误码，0为正常
                example: 0
              data:
                type: object
                properties:
                  msg:
                    type: string
                    example: "已提交/操作成功"

  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header
    petstore_auth:
      type: oauth2
      flows: 
        implicit:
          authorizationUrl: http://example.org/api/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
                